+++ src/AspNetCore.MailKitMailer/Data/MailerContextAbstract.cs
@@ -1,5 +1,6 @@
 ﻿using AspNetCore.MailKitMailer.Domain;
 using AspNetCore.MailKitMailer.Models;
+using MailKit.Net.Smtp;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -46,6 +47,15 @@ namespace AspNetCore.MailKitMailer.Data
         public EmailAddressModel? From { get; set; }
 
         /// <summary>
+        /// Gets or sets the SMTP configuration override.
+        /// When set, this configuration will be used instead of the default SMTP configuration.
+        /// </summary>
+        /// <value>
+        /// The SMTP configuration override, or null to use the default.
+        /// </value>
+        public SMTPConfigModel? SmtpConfigOverride { get; set; }
+
+        /// <summary>
         /// Called when [after send].
         /// </summary>
         /// <param name="serviceProvider">The service provider.</param>
@@ -61,6 +71,15 @@ namespace AspNetCore.MailKitMailer.Data
         public virtual void OnBeforeSend(IServiceProvider serviceProvider)
         {
             
+        }
+
+        /// <summary>
+        /// Called to configure the SMTP client before connecting.
+        /// Override this to modify SMTP client settings per mailer context.
+        /// </summary>
+        /// <param name="client">The SMTP client to configure.</param>
+        public virtual void OnConfigureSmtpClient(SmtpClient client)
+        {
         }
 
         #region HtmlMailHelper